Upgrade PHP and WP
Seems like the image used was a cached one, because 5.6 does not support FSE, and yet I was able to use it no problem. WP 6.2 is the current latest, and PHP 7.4 is not enough for it: there doesn't even seem to be a Docker image for WP 6.2 available with PHP 7.4.